+++ apps/web/src/components/Home/Highlights.tsx
@@ -1,7 +1,6 @@
 import QueuedPublication from '@components/Publication/QueuedPublication';
 import SinglePublication from '@components/Publication/SinglePublication';
 import PublicationsShimmer from '@components/Shared/Shimmer/PublicationsShimmer';
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { CollectionIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
@@ -13,7 +12,7 @@ import { useInView } from 'react-cool-inview';
 import { OptmisticPublicationType } from 'src/enums';
 import { useAppStore } from 'src/store/app';
 import { useTransactionPersistStore } from 'src/store/transaction';
-import { Card } from 'ui';
+import { Card, EmptyState } from 'ui';
 
 const Highlights: FC = () => {
   const currentProfile = useAppStore((state) => state.currentProfile);
