+++ apps/web/src/components/Home/Suggested.tsx
@@ -1,6 +1,5 @@
 import Loader from '@components/Shared/Loader';
 import UserProfile from '@components/Shared/UserProfile';
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { UsersIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
@@ -8,6 +7,7 @@ import type { Profile } from 'lens';
 import { useRecommendedProfilesQuery } from 'lens';
 import type { FC } from 'react';
 import { FollowSource } from 'src/tracking';
+import { EmptyState } from 'ui';
 
 const Suggested: FC = () => {
   const { data, loading, error } = useRecommendedProfilesQuery();
