+++ apps/web/src/components/Nft/NftFeed.tsx
@@ -1,6 +1,5 @@
 import SingleNft from '@components/Nft/SingleNft';
 import NftsShimmer from '@components/Shared/Shimmer/NftsShimmer';
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { CollectionIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
@@ -11,6 +10,7 @@ import type { FC } from 'react';
 import { useState } from 'react';
 import { useInView } from 'react-cool-inview';
 import { CHAIN_ID } from 'src/constants';
+import { EmptyState } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import { mainnet } from 'wagmi/chains';
