+++ apps/web/src/components/Profile/Feed.tsx
@@ -1,6 +1,5 @@
 import SinglePublication from '@components/Publication/SinglePublication';
 import PublicationsShimmer from '@components/Shared/Shimmer/PublicationsShimmer';
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { CollectionIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
@@ -12,7 +11,7 @@ import { useInView } from 'react-cool-inview';
 import { ProfileFeedType } from 'src/enums';
 import { useAppStore } from 'src/store/app';
 import { useProfileFeedStore } from 'src/store/profile-feed';
-import { Card } from 'ui';
+import { Card, EmptyState } from 'ui';
 import formatHandle from 'utils/formatHandle';
 
 interface FeedProps {
