+++ apps/web/src/components/Profile/Following.tsx
@@ -1,6 +1,5 @@
 import Loader from '@components/Shared/Loader';
 import UserProfile from '@components/Shared/UserProfile';
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { UsersIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
@@ -10,6 +9,7 @@ import type { FC } from 'react';
 import { useState } from 'react';
 import { useInView } from 'react-cool-inview';
 import { FollowSource } from 'src/tracking';
+import { EmptyState } from 'ui';
 import formatHandle from 'utils/formatHandle';
 
 interface FollowingProps {
