+++ apps/web/src/components/Profile/NftGallery/Picker.tsx
@@ -1,6 +1,5 @@
 import SingleNft from '@components/Nft/SingleNft';
 import NftPickerShimmer from '@components/Shared/Shimmer/NftPickerShimmer';
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { CheckIcon, CollectionIcon } from '@heroicons/react/outline';
 import { t, Trans } from '@lingui/macro';
@@ -16,6 +15,7 @@ import { CHAIN_ID } from 'src/constants';
 import { useAppStore } from 'src/store/app';
 import type { NftGalleryItem } from 'src/store/nft-gallery';
 import { useNftGalleryStore } from 'src/store/nft-gallery';
+import { EmptyState } from 'ui';
 import formatHandle from 'utils/formatHandle';
 import { mainnet } from 'wagmi/chains';
