+++ apps/web/src/components/Profile/NftGallery/ReviewSelection.tsx
@@ -1,11 +1,11 @@
 import SingleNft from '@components/Nft/SingleNft';
-import { EmptyState } from '@components/UI/EmptyState';
 import { CollectionIcon, XIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
 import type { Nft } from 'lens';
 import React from 'react';
 import type { NftGalleryItem } from 'src/store/nft-gallery';
 import { useNftGalleryStore } from 'src/store/nft-gallery';
+import { EmptyState } from 'ui';
 
 const ReviewSelection = () => {
   const gallery = useNftGalleryStore((state) => state.gallery);
