+++ apps/web/src/components/Shared/Modal/Likes.tsx
@@ -1,5 +1,4 @@
 import UserProfile from '@components/Shared/UserProfile';
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { HeartIcon } from '@heroicons/react/outline';
 import { t } from '@lingui/macro';
@@ -9,6 +8,7 @@ import type { FC } from 'react';
 import { useState } from 'react';
 import { useInView } from 'react-cool-inview';
 import { FollowSource } from 'src/tracking';
+import { EmptyState } from 'ui';
 
 import Loader from '../Loader';
