+++ apps/web/src/components/Shared/Modal/Report/index.tsx
@@ -1,4 +1,3 @@
-import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { Spinner } from '@components/UI/Spinner';
 import { TextArea } from '@components/UI/TextArea';
@@ -13,7 +12,7 @@ import type { FC } from 'react';
 import { useEffect, useState } from 'react';
 import { useGlobalModalStateStore } from 'src/store/modals';
 import { PAGEVIEW, PUBLICATION } from 'src/tracking';
-import { Button, Form, useZodForm } from 'ui';
+import { Button, EmptyState, Form, useZodForm } from 'ui';
 import { object, string } from 'zod';
 
 import Reason from './Reason';
