+++ packages/ui/index.ts
@@ -2,4 +2,5 @@ export { Alert } from './Alert';
 export { Button } from './Button';
 export { Card } from './Card';
 export { Checkbox } from './Checkbox';
+export { EmptyState } from './EmptyState';
 export { FieldError, Form, useZodForm } from './Form';
